refactor(experimental): graphql: group schema typedefs
